Move all schemas to seperate folder.
