+++ pipeline/src/schemas/fitbit_schemas.py
