+++ pipeline/src/schemas/kindle_schemas.py
