+++ pipeline/src/schemas/strong_schemas.py
