+++ pipeline/src/schemas/youtube_schemas.py
