+++ pipeline/src/transformations/fitbit/sleep.py
@@ -5,7 +5,7 @@ import pandas as pd
 import pandera as pa
 from pandera.typing.pandas import DataFrame
 
-from transformations.fitbit.schemas import FitbitSleep, RawFitbitSleep
+from schemas.fitbit_schemas import FitbitSleep, RawFitbitSleep
 from transformations.fitbit.common import extract_json_file_data
 from transformations.utils.io import extract_specific_files_flat
